Split settings-dialog.tsx by tab
The 1121-line file held six tabs' worth of dialog logic, mostly distinct
(profile vs api-keys vs delete-account flows). Reads as one file because
of the shared `extras()` + `SignedOutNotice` helpers, but the tabs
themselves are independent.

Split into a settings-dialog/ folder with index.tsx (shell + nav + section
dispatch) and one file per tab. The two cross-tab helpers live in
shared.tsx. Import path `@/components/settings-dialog` continues to
resolve via the folder's index.tsx — no caller change needed.

Pure extract: no logic, behavior, or markup changed.
