+++ apps/web/src/components/settings-dialog/appearance-tab.tsx
@@ -1,0 +1,40 @@
+import { useTheme } from "@/components/theme-provider"
+import { Button } from "@/components/ui/button"
+import {
+  Field,
+  FieldDescription,
+  FieldGroup,
+  FieldLabel,
+} from "@/components/ui/field"
+
+const THEMES = [
+  { value: "light" as const, label: "Light" },
+  { value: "dark" as const, label: "Dark" },
+  { value: "system" as const, label: "System" },
+]
+
+export function AppearancePanel() {
+  const { theme, setTheme } = useTheme()
+  return (
+    <FieldGroup>
+      <Field>
+        <FieldLabel>Theme</FieldLabel>
+        <FieldDescription>
+          Choose how Arsenyx looks. System follows your OS preference.
+        </FieldDescription>
+        <div className="flex gap-2 pt-1">
+          {THEMES.map((t) => (
+            <Button
+              key={t.value}
+              variant={theme === t.value ? "default" : "outline"}
+              size="sm"
+              onClick={() => setTheme(t.value)}
+            >
+              {t.label}
+            </Button>
+          ))}
+        </div>
+      </Field>
+    </FieldGroup>
+  )
+}
